requirements.txt
urllib3
requests
dnspython
openpyxl